Added integration tests. Tests caught that Team.GoalsPerGame should be a decimal, not an int.

+++ NHL-API/Model/Team.cs
@@ -22,7 +22,7 @@ namespace NHL_API.Model
 
         public int Points { get; set; }
 
-        public int GoalsPerGame { get; set; }
+        public decimal GoalsPerGame { get; set; }
 
         public DateTimeOffset SeasonFirstGameDate { get; set; }
